+++ crates/apollo-compiler/tests/validation/mod.rs
@@ -5,6 +5,7 @@ mod object;
 mod one_of;
 mod operation;
 mod recursion;
+mod september_2025;
 mod types;
 mod variable;